InspIRCd v4: Support real_username (well, faked)
In the 1206 protocol, the UID message now has an extra field for the
real username of the user which is being introduced. We don't track this
and this is unlikely to be useful for other protocols, so we just ignore
it when receiving UID from an InspIRCd 1206 server, and we send the
ident instead of the real_username when sending to a 1206 server.