InspIRCd v4: Fix loop condition in FJOIN memid reset
There was a typo in the InspIRCd v3 code, in the FJOIN handler:
The loop condition said n > 0, while n is never changed in the loop.
It turns out the the loop condition should have been x > 0.

This was fixed for InspIRCd v3 in commit 85716714, but it was
not ported to the InspIRCd v4 protocol support code because
the v4 code was copied from v3 a while ago.

This commit fixes it.